// app/src/redux/action.types.ts
export enum ActionTypes {
  LOGIN_ACTION_START = "LOGIN_ACTION_START",
  SIGNUP_ACTION_START = "SIGNUP_ACTION_START",
  LOGOUT_ACTION_START = "LOGOUT_ACTION_START",
  FETCH_EVENTS_ACTION_START = "FETCH_EVENTS_ACTION_START",

  LOGIN_ACTION_COMPLETE = "LOGIN_ACTION_COMPLETE",
  SIGNUP_ACTION_COMPLETE = "SIGNUP_ACTION_COMPLETE",
  LOGOUT_ACTION_COMPLETE = "LOGOUT_ACTION_COMPLETE",
  FETCH_EVENTS_ACTION_COMPLETE = "FETCH_EVENTS_ACTION_COMPLETE",

  LOGIN_ACTION_ERROR = "LOGIN_ACTION_ERROR",
  SIGNUP_ACTION_ERROR = "SIGNUP_ACTION_ERROR",
  LOGOUT_ACTION_ERROR = "LOGOUT_ACTION_ERROR",
  FETCH_EVENTS_ACTION_ERROR = "FETCH_EVENTS_ACTION_ERROR",

  SET_USER_ACTION = "SET_USER_ACTION",
}